Respect rendered width when adding TUI history (#38075)
## What changed

- Initialize new chat widgets with the current terminal width.
- Determine whether an added history cell is visible using the active history
  render mode and the width available after ambient-pet reservations.
- Saturate diff-summary width calculations so extremely narrow terminals remain
  renderable.

## Testing

- Cover width-sensitive history insertion with an ambient pet, raw-output mode,
  and a diff summary narrower than its indentation.

GitOrigin-RevId: 82d20900e0637431be7646b4f4aba61810ff74c1

+++ codex-rs/tui/src/chatwidget/rendering.rs
@@ -58,7 +58,7 @@ impl ChatWidget {
     }
 
     pub(crate) fn note_rendered_width(&self, width: u16) {
-        self.last_rendered_width.set(Some(width as usize));
+        self.last_rendered_width.set(Some(width));
     }
 }
